requirements.txt
cvxpy-base==1.3.2
et-xmlfile==1.1.0
matplotlib==3.7.2
Mosek==10.0.46
numpy==1.25.2
openpyxl==3.1.2
pandas==2.0.3
pymodbus==3.4.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
tzdata==2023.3
